added loggin
git-svn-id: https://scm.mmbase.org/mmbase/trunk@42611 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ applications/resources/src/main/java/org/mmbase/servlet/HandleServlet.java
@@ -67,7 +67,7 @@ public class HandleServlet extends BridgeServlet {
             // default: one hour
             expires = 60 * 60 * 1000;
         } else {
-            expires = Integer.valueOf(expiresParameter).intValue() * 1000;
+            expires = Long.valueOf(expiresParameter).longValue() * 1000;
         }
 
         String ieCompat = getInitParameter("IECompatibleJpeg");
